refactor: colocate types with their owning modules
Move types from the centralized types.ts into diagnostics.ts, logger.ts,
format.ts, and reporter.ts. Extract generic type utilities into utils.ts.
