+++ src/index.ts
@@ -1,24 +1,29 @@
 export { defineDiagnostics } from './diagnostics'
-
-export { CodedError } from './error'
-export { formatTag, plainFormatter, renderFrame } from './format'
-export { createLogger } from './logger'
-export { consoleReporter, createFetchReporter } from './reporter'
 export type {
   CodeFactory,
-  CreateLoggerOptions,
   DefineDiagnosticsOptions,
   Diagnostic,
-  DiagnosticActions,
   DiagnosticDefinition,
   DiagnosticLevel,
   DiagnosticsMethods,
   DiagnosticsResult,
-  Formatter,
+  Overrides,
+  SourceLocation,
+} from './diagnostics'
+
+export { CodedError } from './error'
+
+export { formatTag, plainFormatter, renderFrame } from './format'
+export type { Formatter } from './format'
+
+export { createLogger } from './logger'
+export type {
+  CreateLoggerOptions,
+  DiagnosticActions,
   Logger,
   LoggerMethods,
   MergeFactories,
-  Overrides,
-  Reporter,
-  SourceLocation,
-} from './types'
+} from './logger'
+
+export { consoleReporter, createFetchReporter } from './reporter'
+export type { Reporter } from './reporter'
